+++ x/fileurl.go
@@ -1,4 +1,4 @@
-package helper
+package x
 
 import (
 	"net/url"
